removed mapdb implementation removed twitter client simplified code
