--- .idea/libraries/Maven__net_jcip_jcip_annotations_1_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: net.jcip:jcip-annotations:1.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/jcip/jcip-annotations/1.0/jcip-annotations-1.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/net/jcip/jcip-annotations/1.0/jcip-annotations-1.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/jcip/jcip-annotations/1.0/jcip-annotations-1.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
