--- .idea/libraries/Maven__net_jpountz_lz4_lz4_1_3_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: net.jpountz.lz4:lz4:1.3.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/jpountz/lz4/lz4/1.3.0/lz4-1.3.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/net/jpountz/lz4/lz4/1.3.0/lz4-1.3.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/jpountz/lz4/lz4/1.3.0/lz4-1.3.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
