--- .idea/libraries/Maven__org_jetbrains_kotlin_kotlin_stdlib_1_0_6.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.jetbrains.kotlin:kotlin-stdlib:1.0.6">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/jetbrains/kotlin/kotlin-stdlib/1.0.6/kotlin-stdlib-1.0.6.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/jetbrains/kotlin/kotlin-stdlib/1.0.6/kotlin-stdlib-1.0.6-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/jetbrains/kotlin/kotlin-stdlib/1.0.6/kotlin-stdlib-1.0.6-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
