--- .idea/libraries/Maven__org_mapdb_elsa_3_0_0_M5.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.mapdb:elsa:3.0.0-M5">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mapdb/elsa/3.0.0-M5/elsa-3.0.0-M5.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mapdb/elsa/3.0.0-M5/elsa-3.0.0-M5-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mapdb/elsa/3.0.0-M5/elsa-3.0.0-M5-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
