--- .idea/libraries/Maven__org_mapdb_mapdb_3_0_3.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.mapdb:mapdb:3.0.3">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mapdb/mapdb/3.0.3/mapdb-3.0.3.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mapdb/mapdb/3.0.3/mapdb-3.0.3-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mapdb/mapdb/3.0.3/mapdb-3.0.3-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
