--- src/main/java/com/zuhlke/ta/prototype/solutions/common/TweetStore.java
@@ -1,11 +1,0 @@
-package com.zuhlke.ta.prototype.solutions.common;
-
-import com.zuhlke.ta.prototype.Tweet;
-
-import java.util.stream.Stream;
-
-public interface TweetStore {
-    void importTweets(Stream<Tweet> tweets);
-
-    Stream<Tweet> tweets();
-}
