--- src/main/java/com/zuhlke/ta/prototype/solutions/mapdb/MapDBTweetService.java
@@ -1,10 +1,0 @@
-package com.zuhlke.ta.prototype.solutions.mapdb;
-
-import com.zuhlke.ta.prototype.SentimentAnalyzer;
-import com.zuhlke.ta.prototype.solutions.common.PersistentTweetService;
-
-public class MapDBTweetService extends PersistentTweetService{
-    public MapDBTweetService(SentimentAnalyzer sentimentAnalyzer) {
-        super(sentimentAnalyzer, new MapDBTweetStore());
-    }
-}
