+++ src/main/resources/spark/template/freemarker/index.html
@@ -1,7 +1,7 @@
 <html>
     <head>
         <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-        <meta http-equiv="refresh" content="10" >
+        <meta http-equiv="refresh" content="15" >
     </head>
     <body>
         <h1>Twitter Analytics</h1>
